+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 + Add `with_maximum` option to `ValidateLengthMatcher`.
 + Add `with_max` and `is_at_most` aliases to `ValidateLengthMatcher#with_maximum`.
 + Add `within` option to `ValidateLengthMatcher`.
++ Add `in` alias to `ValidateLengthMatcher#within`.